Num2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -1,8 +1,9 @@
 package lang.print.gaps.task5;
 
 public class NumberDivider {
-    void divide(int toBeDivided) {
-        int Raz =toBeDivided/5;
+    void divide(double toBeDivided) {
+
+        double Raz =toBeDivided/5;
         System.out.println(Raz);
     }
 }
